Fixed some small details for showing Popup detailing that quantity received is more or less than pre re quantity - ReceivingProductsDataClasses.kt , ReceivingProductsDetailsFragment.kt, Services.kt, ReceivingProductsViewModel.kt

+++ app/src/main/java/com/scannerapp/cdihandheldscannerviewactivity/Utils/Network/Services.kt
@@ -289,7 +289,7 @@ interface ReceivingProductsServices {
     @GET("validateQuantityAgainstPreReceiving")
     suspend fun validateQuantityAgainstPreReceiving(
         @Query("itemNumber") itemNumber: String,
-        @Query("quantityEntered") quantityEntered: Int,
+        @Query("quantityEntered") quantityEntered: Float,
         @Query("receivingNumber") receivingNumber: String,
         @Query("warehouseNumber") warehouseNumber: Int,
         @Query("companyID") companyID: String
